server/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
pandas==2.0.1
mock==5.0.2
django-extensions==3.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-factory-boy==1.0.0
factory-boy==3.3.0
django-filter==23.3